+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `payload_id` was not passed properly to `StateDB` for logging, via the state
   change decorator. Updated the decorator to be a Descriptor, which may make
   type checking happier. ([#270])
+- added function `cirrus.lib2.utils.cold_start` to move overhead of boto
+  client/resource cache instantiation out of lambda function execution. ([#272])
 
 ### Added
 
@@ -926,6 +928,7 @@ Initial release
 [#266]: https://github.com/cirrus-geo/cirrus-geo/pull/266
 [#267]: https://github.com/cirrus-geo/cirrus-geo/pull/267
 [#270]: https://github.com/cirrus-geo/cirrus-geo/pull/270
+[#272]: https://github.com/cirrus-geo/cirrus-geo/pull/272
 [f25acd4]: https://github.com/cirrus-geo/cirrus-geo/commit/f25acd4f43e2d8e766ff8b2c3c5a54606b1746f2
 [85464f5]: https://github.com/cirrus-geo/cirrus-geo/commit/85464f5a7cb3ef82bc93f6f1314e98b4af6ff6c1
 [1b89611]: https://github.com/cirrus-geo/cirrus-geo/commit/1b89611125e2fa852554951343731d1682dd3c4c